pyproject.toml
[tool.poetry]
name = "rdt"
version = "0.1.0"
description = "Robotics Dev Tools — A unified, recipe-driven CLI for build, test, CI/CD, and demo workflows."
authors = [
    "Thibault Poignonec <thibault.poignonec@gmail.com>",
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/tpoignonec/ros2-devtools"
keywords = ["robotics", "devops", "cli", "ros2", "ci", "dagger"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{ include = "rdt", from = "src" }]
include = ["src/rdt/recipes/ros2/templates/*"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1"
pydantic = "^2.0"
rich = "^13.0"
pyyaml = "^6.0"

[tool.poetry.extras]
ci = ["dagger-io", "anyio"]
demo = ["rocker", "etherlab-rocker"]
all = ["dagger-io", "anyio", "rocker", "etherlab-rocker"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
ruff = ">=0.4"
mypy = "^1.10"
dagger-io = ">=0.9,<1.0"
anyio = ">=4.0"

[tool.poetry.scripts]
rdt = "rdt.cli:cli"

[tool.poetry.plugins."rdt.recipes"]
ros2 = "rdt.recipes.ros2:ROS2Recipe"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# ── Ruff ──────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

# ── Mypy ──────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true

# ── Pytest ────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
